Create readme.md
# 3key

![3KEY](https://i.redd.it/sl6apr271fg71.jpg)

Keymap for a Scramble turned on its side with an OLED and encoder.

Features
* Encoder changes layers
* Audio layer has an encoder volume toggle

VIA can be enabled in rules.mk.  With VIA enabled, create your OLED layer displays in QMK and your key layout and macros in VIA.

DFU mode is accessible from the base QMK layer using the encoder click, or from the Admin/DFU layer with the F3 key. (according to photo)